requirements.txt
-i https://pypi.org/simple
asgiref==3.6.0 ; python_version >= '3.7'
asttokens==2.2.1
backcall==0.2.0
backports.zoneinfo==0.2.1 ; python_version < '3.9'
decorator==5.1.1 ; python_version >= '3.5'
django==4.1.6
executing==1.2.0
ipython==8.9.0
jedi==0.18.2 ; python_version >= '3.6'
matplotlib-inline==0.1.6 ; python_version >= '3.5'
parso==0.8.3 ; python_version >= '3.6'
pexpect==4.8.0 ; sys_platform != 'win32'
pickleshare==0.7.5
prompt-toolkit==3.0.36 ; python_full_version >= '3.6.2'
ptyprocess==0.7.0
pure-eval==0.2.2
pygments==2.14.0 ; python_version >= '3.6'
python-dateutil==2.8.2
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.3 ; python_version >= '3.5'
stack-data==0.6.2
traitlets==5.9.0 ; python_version >= '3.7'
wcwidth==0.2.6
